refactor: Make empty environment validation method less ambiguous
Previously, the method to check if the environment is empty was
called "validateEnvironmentRoot". This commit changes that to
"isEnvironmentEmpty" in preparation for another method that is
called "isUnistackEnviroment", which will check if the environment
is a valid unistack project. This commit also refactors the
previous unit tests for the "isEnvironmentEmpty" method.