test: Ensure message transport unit tests are atomic
Previously, the tests were bleeding into each other due to the
structure of the message transport and transport classes, as well
as failures to properly mock side effects. This commit changes the
class structure and uses static methods to achieve the same goal,
and ensures that all side effects are properly mocked.