test: Add ability to specify a single test suite for analysis
Previously, one had to run all the specs included in the "--quick"
option, and that was the fastest way to debug while maintaining
the integrity of the tests. As the specs grow, this becomes a nuance.
This commit aliviates that problem by adding the ability to specify
a single test suite.