refactor: Fully embrace the Redux architecture in unistack's core
Previously, the system was on a loosely defined architecture
inspired by Redux, and was similar in a number of aspects. After
having done an analysis regarding future considerations, both
for the app environment and the unistack core, the following
reasons for embracing Redux stood out:

1) Open source contributors will likely have an easier time
understanding the system on their own, pontentially increasing
contributions.
2) It has been tried and tested by a fairly large community.
3) It'll make it easier for me to explain the system to others.
4) The backend doesn't actually need the MVC architecture. I had
originally thought of a way to merge Redux philosophies with
the MVC paradigm, but the more I worked out the details the more
I realized that we just don't need it.

This commit essentially rewrites/reorganizes everything to make way
for a solid Redux future.