feat: Ensure UniStack is in some end-to-end functioning state.
Previously, we were in the mist of an architecture switch, which
prompted me to reanalyze some fundamentals, which led to some
further experiments, which then allowed me to come to a few
realizations, which I intend to do a write up on soon. This, and a
promise to myself that I would have some sort of "beta" offering out
by October 1st, has kept me from reimplementing every feature
contained in the old architecture. Regardless, we have arrived at a
very reasonable point, at the bare minimun, UniStack is useful for
prototyping. Overall, I'm glad this direction was taken.